requirements.txt
python-suanpan-slim
pyside6
